+++ tests/headers/typedef_same_name.h
@@ -1,0 +1,3 @@
+typedef struct SameS SameS;
+typedef enum SameE SameE;
+typedef union SameU SameU;