+++ tests/tests.rs
@@ -15,3 +15,4 @@ mod test_struct;
 mod test_union;
 mod test_builtins;
 mod test_ints;
+mod test_typedef;