+++ Changelog.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   in the output of `--help`)
 - Map stdint.h + size_t types to Rust native ones (#256)
 - Default to fail on unknown type (see `-allow-unknown-types`)
+- Convert C `typedef struct {} Test` to rust `struct Test {}` (#169)
 
 ### Added
 - `-no-rust-enums` generate integer constants instead of enums